@@ -1,6 +1,7 @@
 package com.gp.chat.model
 
 data class NetworkChatGroup(
-    val name : String="" ,
+    val name : String="",
+    val picURL: String="",
     val members :Map<String,Boolean> = emptyMap(),
 )
