@@ -1,7 +1,9 @@
 package com.gp.users.model
 
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
 import java.util.Date
-
+@Parcelize
 data class User(
     val firstName: String="",
     val lastName: String="",
@@ -11,4 +13,4 @@ data class User(
     val birthdate: Date = Date(),
     val bio: String="",
     val administration: String="",
-)
+): Parcelable
