@@ -2,10 +2,7 @@ package com.gp.chat.adapter
 
 import android.content.Context
 import android.graphics.drawable.Drawable
-import android.view.ViewGroup
-import android.widget.Button
 import android.widget.ImageView
-import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
@@ -13,13 +10,10 @@ import com.bumptech.glide.request.target.CustomTarget
 import com.bumptech.glide.request.transition.Transition
 import com.google.android.material.chip.Chip
 import com.google.android.material.chip.ChipGroup
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
 import com.gp.chat.R
 import com.gp.chat.listener.OnGroupMembersChangeListener
-import com.gp.chat.model.RecentChat
 import com.gp.users.model.User
-private val currentUser = Firebase.auth.currentUser
+
 @BindingAdapter("chat:imageUrl")
 fun setProfilePicture(view: ImageView, picUrl: String?) {
     if (picUrl != null) {
@@ -37,7 +31,7 @@ fun setSelectedMembers(view: ChipGroup, selectedMembers: List<User>, context: Co
     view.removeAllViews()
     if (view.childCount==0) {
         selectedMembers.forEach {user ->
-            val label = "${user.firstName} ${user.lastName}"
+            val label = user.firstName
             val chip = Chip(context)
             chip.text = label
             chip.textSize = 14f
