@@ -5,6 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.gp.chat.repository.MessageRepository
 import com.gp.socialapp.utils.State
+import com.gp.users.model.SelectableUser
+import com.gp.users.model.SelectableUser.Companion.toSelectableUser
 import com.gp.users.model.User
 import com.gp.users.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -21,7 +23,7 @@ class CreateGroupChatViewModel @Inject constructor(
     private val chatRepo: MessageRepository
 ) : ViewModel() {
     val uiState = MutableStateFlow(CreateGroupChatUiState())
-    private val _users = MutableStateFlow(emptyList<User>())
+    private val _users = MutableStateFlow(emptyList<SelectableUser>())
     val users = _users.asStateFlow()
     private val currentUserEmail = userRepo.getCurrentUserEmail()
     init {
@@ -30,26 +32,44 @@ class CreateGroupChatViewModel @Inject constructor(
     }
 
     fun addMember(user: User) {
-        val updatedMembers = uiState.value.selectedMembers.toMutableList()
-        updatedMembers.add(user)
-        uiState.value = uiState.value.copy(selectedMembers = updatedMembers.toList())
+        viewModelScope.launch (Dispatchers.Default){
+            val updatedMembers = uiState.value.selectedMembers.toMutableList()
+            updatedMembers.add(user)
+            _users.value = _users.value.map {
+                if (it.data.email == user.email) {
+                    Log.d("seerde", "User status before edit in viewmodel: ${it.isSelected}")
+                    it.copy(isSelected = true)
+                } else {
+                    it
+                }
+            }
+            Log.d("seerde", "User status before edit in viewmodel: ${_users.value.find { it.data.email == user.email }!!.isSelected}")
+            uiState.value = uiState.value.copy(selectedMembers = updatedMembers.toList())
+        }
     }
 
     fun removeMember(user: User) {
-        val updatedMembers = uiState.value.selectedMembers.toMutableList()
-        updatedMembers.remove(user)
-        uiState.value = uiState.value.copy(selectedMembers = updatedMembers.toList())
+        viewModelScope.launch(Dispatchers.Default){
+            val updatedMembers = uiState.value.selectedMembers.toMutableList()
+            updatedMembers.remove(user)
+            _users.value = _users.value.map {
+                if (it.data.email == user.email) {
+                    it.copy(isSelected = false)
+                } else {
+                    it
+                }
+            }
+            uiState.value = uiState.value.copy(selectedMembers = updatedMembers.toList())
+        }
     }
 
     fun getListOfUsers() {
         viewModelScope.launch(Dispatchers.IO) {
-            Log.d("EDREES", "GetListOfUsers Called")
-            Log.d("EDREES", "Before fetchUsers()")
             userRepo.fetchUsers().collect {
                 when (it) {
                     is State.SuccessWithData -> {
-                        _users.value = it.data.filter { it.email != currentUserEmail }
-                        Log.d("EDREES", "User: ${_users.value.map { it.email }}")
+                        _users.value = it.data.map{it.toSelectableUser()}.filter { it.data.email != currentUserEmail }
+                        Log.d("EDREES", "User: ${_users.value.map { it.data.email }}")
                     }
 
                     is State.Error -> {
