@@ -1,5 +1,6 @@
 package com.gp.chat.presentation.home
 
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -7,6 +8,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.appcompat.widget.PopupMenu
 import androidx.appcompat.widget.Toolbar
 import androidx.fragment.app.Fragment
@@ -29,6 +31,9 @@ import com.gp.chat.model.RecentChat
 import com.gp.socialapp.database.model.UserEntity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
+import java.time.ZonedDateTime
+import java.time.format.DateTimeFormatter
+import java.util.Locale
 
 @AndroidEntryPoint
 class ChatHome : Fragment(), OnRecentChatClicked {
@@ -47,6 +52,7 @@ class ChatHome : Fragment(), OnRecentChatClicked {
     }
 
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val toolbar = requireActivity().findViewById<Toolbar>(R.id.toolbar)
@@ -57,21 +63,31 @@ class ChatHome : Fragment(), OnRecentChatClicked {
         recyclerView.adapter = chatAdapter
 
         lifecycleScope.launch {
-            viewModel.recentChats.flowWithLifecycle(lifecycle).collect {
-                chatAdapter.submitList(it)
-            }
-        }
+            val  formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss z", Locale.ENGLISH)
 
+            viewModel.recentChats.flowWithLifecycle(lifecycle).collect { it ->
+                chatAdapter.submitList(it.map {
+                    it.copy(
+                        timestamp = DateTimeFormatter
+                            .ofPattern("hh:mm", Locale.ENGLISH)
+                            .format(ZonedDateTime.parse(it.timestamp,formatter))
+                    )
+                })
+            }
+
+
+        }
         floatingActionButton.setOnClickListener {
             val action = ChatHomeDirections.actionChatHomeToNewChat()
             findNavController().navigate(action)
         }
+
     }
 
     override fun onRecentChatClicked(
         recentChat: RecentChat
     ) {
-        with(recentChat){
+        with(recentChat) {
             val action = if (recentChat.privateChat) {
                 ChatHomeDirections.actionChatHomeToPrivateChatFragment(
                     chatId = id,
@@ -91,9 +107,9 @@ class ChatHome : Fragment(), OnRecentChatClicked {
         }
 
 
-
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun leaveGroup(groupId: String) {
         viewModel.leaveGroup(groupId)
     }
